fix: rename preview modal
